requirements.txt
aiofiles==24.1.0
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
antiorm==1.2.1
anyio==4.3.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==2.4.1
async-timeout==4.0.3
asyncpg==0.30.0
attrs==23.2.0
azure-core==1.32.0
azure-storage-blob==12.25.0
backoff==2.2.1
bcrypt==4.1.2
blinker==1.7.0
blis==1.2.1
build==1.2.1
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
cloudpathlib==0.21.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
contourpy==1.3.2
cryptography==42.0.5
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.4
datasets==2.20.0
db==0.1.1
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.110.1
fastapi-cli==0.0.4
filelock==3.13.4
flatbuffers==24.3.25
fonttools==4.58.0
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.24.1
google-auth==2.29.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.1
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.70.0
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.7
importlib-metadata==7.0.0
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==8.18.1
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jiter==0.7.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kaleido==0.2.1
kiwisolver==1.4.8
kubernetes==30.1.0
langchain==0.1.15
langchain-community==0.0.32
langchain-core==0.1.41
langchain-openai==0.1.2
langchain-text-splitters==0.0.1
langcodes==3.5.0
langdetect==1.0.9
langsmith==0.1.45
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.10.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.13
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.1
openai==1.17.0
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.0
overrides==7.7.0
packaging==23.2
pandas==2.2.2
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.0
plotly==5.22.0
posthog==3.5.0
preshed==3.0.9
prompt-toolkit==3.0.43
proto-plus==1.26.0
protobuf==4.25.3
psutil==5.9.8
psycopg2==2.9.9
pulsar-client==3.4.0
pure-eval==0.2.2
pyarrow==15.0.2
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
pyodbc==5.2.0
pyOpenSSL==24.1.0
pyparsing==3.2.3
PyPika==0.48.9
pyproject_hooks==1.0.0
pyreadline3==3.4.1
pysbd==0.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
ragas==0.0.22
referencing==0.34.0
regex==2023.12.25
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
shellingham==1.5.4
six==1.16.0
smart-open==7.1.0
smmap==5.0.1
sniffio==1.3.1
snowflake-connector-python==3.11.0
snowflake-sqlalchemy==1.5.1
sortedcontainers==2.4.0
spacy==3.8.7

pysqlite3-binary


spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.4
SQLAlchemy==1.4.52
srsly==2.5.1
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
tenacity==8.2.3
thinc==8.3.4
tiktoken==0.6.0
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.4
toolz==0.12.1
tornado==6.4
tqdm==4.66.5
traitlets==5.14.2
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
urllib3==1.26.19
uvicorn==0.29.0
wasabi==1.1.3
watchdog==4.0.0
watchfiles==0.21.0
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.7.0
websockets==12.0
wordcloud==1.9.4
wrapt==1.16.0
XlsxWriter==3.2.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.1